pyproject.toml
[tool.poetry]
    name = "analysis-model"
    version = "0.1"
    description = ""
    authors = ["huang0520 <huangmorris4x@gmail.com>"]
    readme = "README.md"

    [tool.poetry.dependencies]
        python = ">=3.12, < 3.13"
        PyYAML = "^6.0.1"
        result = "^0.16.1"
        tqdm = "^4.66.4"
        numpy = "^1.26.4"
        pandas = "^2.2.2"
        schema = "^0.7.7"
        attrs = "^23.2.0"
        cattrs = "^23.2.3"

    [tool.poetry.group.dev.dependencies]
        icecream = "^2.1.3"
        notebook = "^7.1.3"

[build-system]
    requires = ["poetry-core"]
    build-backend = "poetry.core.masonry.api"

[tool.ruff]
    target-version = "py312"
    preview = true

    select = [
        "F", # Pyflakes
        "E", # Pydocstyle Errors
        "W", # Pydocstyle Warnings
        "C90", # McCabe Complexity
        "I001", # isort
        "N", # pep8-naming
        "UP", # pyupgrade
        "PD", # Pandas-vet
        "FLY", # Flynt
        "RUF", # Ruff
        "NPY", # Numpy
        "PERF", # Performance
        "PL", # Pylint

        # Flake8
        # "PTH", # Using pathlib
        "ISC", # Implicit string concatenation
        "C4", # Comprehensive
        "ANN", # Annotations
        "BLE", # No blind except
        "ICN", # Import conventions
        "A", # Avoid shadowing builtins
        "B", # Bugbear
        "EM", # Error messaging
        "LOG", # Logging
        "G", # Logging format
        "TID", # Tidy imports
        "SIM", # Simplicity
    ]
    ignore = ["F401", "E402", "E203", "ANN401"]
    include = ["*.py", "*.pyi", "**/pyproject.toml", "*.ipynb"]

    line-length = 88

    [tool.ruff.pydocstyle]
        convention = "google" # Accepts: "google", "numpy", or "pep257".
